stores/library/library_test.go
package library

import (
	"context"
	"testing"

	"gofr.dev/ZopSmart_GoFr_Project/models"
	"gofr.dev/pkg/datastore"
	"gofr.dev/pkg/gofr"

	"github.com/stretchr/testify/assert"
)

func initializeTest(t *testing.T) *gofr.Context {
	// Your initialization code for the library use case here
}

func TestCustomer_Get(t *testing.T) {
	tests := []struct {
		desc string
		name string
		resp models.Book
		err  error
	}{
		{
			desc: "get existing book",
			name: "12345", // Replace with an existing book ID
			resp: models.Book{
				ID:       "12345",
				Title:    "Sample Book Title",
				Author:   "John Doe",
				Publisher: "Acme Publishing",
			},
			err: nil,
		},
		// Add more test cases as needed
	}

	store := New() // Replace with the actual initialization of the library store
	ctx := initializeTest(t)

	for i, tc := range tests {
		resp, err := store.Get(ctx, tc.name)

		assert.Equal(t, tc.err, err, "TEST[%d] %s: failed", i, tc.desc)
		assert.Equal(t, tc.resp, resp, "TEST[%d] %s: failed", i, tc.desc)
	}
}

func TestModel_Create(t *testing.T) {
	tests := []struct {
		desc     string
		book     models.Book
		err      error
	}{
		{
			desc: "create new book",
			book: models.Book{
				ID:       "67890", // Replace with a new book ID
				Title:    "New Book Title",
				Author:   "Jane Smith",
				Publisher: "XYZ Publications",
			},
			err: nil,
		},
		// Add more test cases as needed
	}

	store := New() // Replace with the actual initialization of the library store
	ctx := initializeTest(t)

	for i, tc := range tests {
		err := store.Create(ctx, tc.book)

		assert.Equal(t, tc.err, err, "TEST[%d] %s: failed", i, tc.desc)
	}
}

func TestModel_Delete(t *testing.T) {
	tests := []struct {
		desc  string
		name  string
		count int
		err   error
	}{
		{
			desc:  "delete existing book",
			name:  "12345", // Replace with an existing book ID
			count: 1,
			err:   nil,
		},
		// Add more test cases as needed
	}

	store := New() // Replace with the actual initialization of the library store
	ctx := initializeTest(t)

	for i, tc := range tests {
		count, err := store.Delete(ctx, tc.name)

		assert.Equal(t, tc.err, err, "TEST[%d] %s: failed", i, tc.desc)
		assert.Equal(t, tc.count, count, "TEST[%d] %s: failed", i, tc.desc)
	}
}
